Initial refactoring -  split code into 4 files
