+++ .idea/misc.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.6.1 (C:\Users\Tanvi Gupta\Anaconda3\python.exe)" project-jdk-type="Python SDK" />
+</project>
